move range_categories() function to distances.py, add some tests
