+++ co2calculator/calculate.py
@@ -19,9 +19,8 @@ from .constants import (
     FerryClass,
     ElectricityFuel,
     TransportationMode,
-    RangeCategory,
 )
-from .distances import create_distance_request, get_distance
+from .distances import create_distance_request, get_distance, range_categories
 
 script_path = str(Path(__file__).parent)
 emission_factor_df = pd.read_csv(f"{script_path}/../data/emission_factors.csv")
@@ -485,31 +484,6 @@ def calc_co2_businesstrip(
     range_category, range_description = range_categories(distance)
 
     return emissions, distance, range_category, range_description
-
-
-def range_categories(distance: Kilometer) -> Tuple[RangeCategory, str]:
-    """Function to categorize a trip according to the travelled distance
-
-    :param distance: Distance travelled in km
-    :type distance: Kilometer
-    :return: Range category of the trip [very short haul, short haul, medium haul, long haul]
-             Range description (i.e., what range of distances does to category correspond to)
-    :rtype: tuple[RangeCategory, str]
-    """
-    if distance <= 500:
-        range_cat = RangeCategory.VERY_SHORT_HAUL
-        range_description = "below 500 km"
-    elif distance <= 1500:
-        range_cat = RangeCategory.SHORT_HAUL
-        range_description = "500 to 1500 km"
-    elif distance <= 4000:
-        range_cat = RangeCategory.MEDIUM_HAUL
-        range_description = "1500 to 4000 km"
-    else:
-        range_cat = RangeCategory.LONG_HAUL
-        range_description = "above 4000 km"
-
-    return range_cat, range_description
 
 
 def get_emission_factor(
