+++ tests/unit/test_calculate.py
@@ -86,8 +86,7 @@ def test_calc_co2_motorbike(
         pytest.param(10, None, None, 50, None, 0.39, id="occupancy: 50"),
         pytest.param(10, None, None, 80, None, 0.26, id="occupancy: 80"),
         pytest.param(10, None, None, 100, None, 0.22, id="occupancy: 100"),
-        # NOTE: 'local' as `vehicle_range` fails with IndexError
-        # pytest.param(10, None, None, None, "local", 21.63, id="vehicle_range: 'local'"),
+        pytest.param(10, None, None, None, "local", 0.39, id="vehicle_range: 'local'"),
         pytest.param(
             10,
             None,
@@ -96,6 +95,33 @@ def test_calc_co2_motorbike(
             "long-distance",
             0.39,
             id="vehicle_range: 'long-distance'",
+        ),
+        pytest.param(
+            10,
+            "small",
+            "diesel",
+            None,
+            "long-distance",
+            0.39,
+            id="size: 'small', fuel_type: `diesel`, vehicle_range: 'long-distance'",
+        ),
+        pytest.param(
+            10,
+            "medium",
+            "cng",
+            None,
+            "long-distance",
+            0.62,
+            id="fuel_type: `cng` and size",
+        ),
+        pytest.param(
+            10,
+            "small",
+            "hydrogen",
+            None,
+            "local",
+            0.25,
+            id="fuel_type: `hydrogen` and size",
         ),
     ],
 )
@@ -206,9 +232,8 @@ def test_calc_co2_plane__invalid_distance_seating_combo() -> None:
     [
         pytest.param(None, 11.29, id="defaults"),
         pytest.param("average", 11.29, id="seating_class: 'average'"),
-        # TODO: IndexError for commented test arguments
-        # pytest.param("Foot passenger", 1, id="seating_class: 'Foot passenger'"),
-        # pytest.param("Car passenger", 1, id="seating_class: 'Car passenger"),
+        pytest.param("foot_passenger", 1.87, id="seating_class: 'Foot passenger'"),
+        pytest.param("car_passenger", 12.95, id="seating_class: 'Car passenger"),
     ],
 )
 def test_calc_ferry(seating_class: Optional[str], expected_emissions: float) -> None:
@@ -304,9 +329,6 @@ def test_commuting(
         pytest.param(4000, "medium_haul", "1500 to 4000 km", id="Distance: 4000 km"),
         pytest.param(4001, "long_haul", "above 4000 km", id="Distance: 4001 km"),
         pytest.param(42.7, "very_short_haul", "below 500 km", id="float"),
-        # NOTE: For the time being signed values are possible
-        pytest.param(-42.7, "very_short_haul", "below 500 km", id="signed float"),
-        pytest.param(-500, "very_short_haul", "below 500 km", id="signed int"),
     ],
 )
 def test_range_categories(
@@ -319,6 +341,14 @@ def test_range_categories(
 
     assert actual_category == expected_category
     assert actual_description == expected_description
+
+
+def test_range_categories_negative_distance():
+    """Test: Categorization of ranges when using negative distance
+    Expect: Test fails
+    """
+    with pytest.raises(ValueError):
+        candidate.range_categories(-20)
 
 
 @pytest.mark.parametrize(
